[c04] Open file for c04
Oy. All the examples of File I/O for Rust makes use of the try! macro,
but that can't be directly used in main() (or any void function) because
of an implict return in it.

Instead, anything with try! must be done in other functions. Doing the
Result<()> is apparently standard Rust practice for "I don't care about
an actual return value, just about returning the error"